Modifiche al sistema delle policy

+++ app/Http/Controllers/SandwichController.php
@@ -4,10 +4,23 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\SandwichRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Sandwich;
 
 class SandwichController extends Controller
 {
+    private $logged_user;
+    private $user_id;
+
+    public function __construct(Request $request){
+        // Utente che fa la chiamata api
+        $this->logged_user = $request->user('api');
+        // Id dell'utente associato all'ordine
+        if ($request->has('api_token')) {
+            $this->user_id = in_array($request->user('api')->role, array('bar')) ? $request->user_id : $request->user('api')->id;             
+        }
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -16,7 +29,6 @@ class SandwichController extends Controller
     public function index()
     {
         $sandwiches = Sandwich::all();
-
         return view('/sandwiches/index')->with(compact('sandwiches'));
     }
 
@@ -28,7 +40,6 @@ class SandwichController extends Controller
     public function index_api()
     {
         $sandwiches = Sandwich::all();
-
         return $sandwiches;
     }
 
@@ -55,16 +66,23 @@ class SandwichController extends Controller
 
     public function store_api(SandwichRequest $request)
     {
-        //$this->authorize('create', User::class);
-        try {
-            $sandwich = new Sandwich;
-            $sandwich->name = $request->name;
-            $sandwich->price = $request->price;
-            $sandwich->description = $request->description;
-            $sandwich->save();
-            return response()->json(['message' => 'Panino creato.'], 200);
-        } catch (\Throwable $th) {
-            return response()->json(['message' => "Errore nell'operazione. Ricontrollare i dati inseriti (il panino potrebbe essere già presente)."], 200);
+        if ($this->logged_user->can('create', Sandwich::class)) {
+            if(!app(\App\Http\Controllers\OrderController::class)->checkRetireTime()){
+                try {
+                    $sandwich = new Sandwich;
+                    $sandwich->name = $request->name;
+                    $sandwich->price = $request->price;
+                    $sandwich->description = $request->description;
+                    $sandwich->save();
+                    return response()->json(['message' => 'Panino creato.'], 200);
+                } catch (\Throwable $th) {
+                    return response()->json(['message' => "Errore nell'operazione. Ricontrollare i dati inseriti (il panino potrebbe essere già presente)."], 400);
+                }
+            }else{
+                return response()->json(['message' => "Errore nell'operazione. Attendere l'orario di ritiro per effettuare cambiamenti nel listino."], 403);
+            }
+        }else{
+            return response()->json(['message' => "Errore nell'operazione. Utente non autorizzato."], 401);
         }
     }
 
@@ -103,7 +121,7 @@ class SandwichController extends Controller
         //
     }
 
-    /**
+    /** 
      * Remove the specified resource from storage.
      *
      * @param  int  $id
@@ -123,9 +141,16 @@ class SandwichController extends Controller
      */
     public function soft_destroy_api($id)
     {
-        //$this->authorize('delete', Sandwich::class);
-        $sandwich = Sandwich::findOrFail($id);
-        $sandwich->delete();
-        return response()->json(['message' => 'Eliminazione effettuata.'], 200);
+        if ($this->logged_user->can('create', Sandwich::class)) {
+            if(!app(\App\Http\Controllers\OrderController::class)->checkRetireTime()){   
+                $sandwich = Sandwich::findOrFail($id);
+                $sandwich->delete();
+                return response()->json(['message' => 'Eliminazione effettuata.'], 200);
+            }else{
+                return response()->json(['message' => "Errore nell'operazione. Attendere l'orario di ritiro per effettuare cambiamenti nel listino."], 403);
+            }
+        }else{
+            return response()->json(['message' => "Errore nell'operazione. Utente non autorizzato."], 401);
+        }
     }
 }
